server/requirements.txt
fastapi
uvicorn
supabase
alpaca-trade-api
pydantic
pycryptodome
pyyaml
python-dotenv
